Create 366FindLeavesofBinaryTree
For each node, we decide the level which is nothing but a number denoting in which iteration will this node be removed. And this is based on the fact that what is the maximum level of its children because a node cant be removed until all its children have been removed.
Lev(Node) = max(Lev(Children) + 1, the extra 1 is there because it will be removed in the next iteration after all its children have been removed.